Moved sync method to table API
